Created router for health check endpoint
